requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
certifi==2025.11.12
distro==1.9.0
exceptiongroup==1.3.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
magic-filter==1.0.12
multidict==6.7.0
openai==2.8.1
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
